+++ addOns/openapi/src/main/java/org/zaproxy/zap/extension/openapi/converter/swagger/SwaggerConverter.java
@@ -39,17 +39,22 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
+import java.util.regex.Pattern;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
+import org.parosproxy.paros.model.Model;
 import org.zaproxy.zap.extension.openapi.converter.Converter;
 import org.zaproxy.zap.extension.openapi.generators.Generators;
 import org.zaproxy.zap.extension.openapi.network.RequestModel;
+import org.zaproxy.zap.model.Context;
+import org.zaproxy.zap.model.StructuralNodeModifier;
 import org.zaproxy.zap.model.ValueGenerator;
 
 public class SwaggerConverter implements Converter {
@@ -58,6 +63,7 @@ public class SwaggerConverter implements Converter {
     private static final String BASE_KEY_I18N = "openapi.swaggerconverter.";
 
     private static Logger LOG = LogManager.getLogger(SwaggerConverter.class);
+    private static final Pattern PATH_PART_PATTERN = Pattern.compile("\\{.*}");
     private final UriBuilder targetUriBuilder;
     private final UriBuilder definitionUriBuilder;
     private String defn;
@@ -65,6 +71,8 @@ public class SwaggerConverter implements Converter {
     private RequestModelConverter requestConverter;
     private Generators generators;
     private List<String> errors = new ArrayList<>();
+    private Set<String> apiUrls;
+    private OpenAPI openAPI;
 
     public SwaggerConverter(String defn, ValueGenerator valGen) {
         this(null, null, defn, valGen);
@@ -170,7 +178,7 @@ public class SwaggerConverter implements Converter {
     }
 
     private List<OperationModel> readOpenAPISpec() throws SwaggerException {
-        OpenAPI openAPI = getOpenAPI();
+        openAPI = getOpenAPI();
         if (openAPI == null) {
             throw new SwaggerException(
                     Constant.messages.getString(BASE_KEY_I18N + "parse.defn.exception", defn));
@@ -179,8 +187,8 @@ public class SwaggerConverter implements Converter {
         List<OperationModel> operations = new ArrayList<>();
         List<UriBuilder> serverUriBuilders =
                 createUriBuilders(openAPI.getServers(), definitionUriBuilder);
-        for (String url :
-                createApiUrls(serverUriBuilders, targetUriBuilder, definitionUriBuilder)) {
+        apiUrls = createApiUrls(serverUriBuilders, targetUriBuilder, definitionUriBuilder);
+        for (String url : apiUrls) {
             addOperations(openAPI, url, operations);
         }
         return operations;
@@ -408,5 +416,105 @@ public class SwaggerConverter implements Converter {
 
         swaggerParseResult.setMessages(errors);
         return swaggerParseResult;
+    }
+
+    private List<StructuralNodeModifier> getStructuralNodeModifiers() {
+        if (openAPI == null) {
+            throw new IllegalStateException("The OpenAPI was not yet imported.");
+        }
+
+        String targetUrl = apiUrls.stream().findFirst().get();
+
+        Set<String> ddnRegexStrings = new HashSet<>();
+        // check for DDN on each key (path) in the spec
+        for (Map.Entry<String, PathItem> entry : openAPI.getPaths().entrySet()) {
+            String key = entry.getKey();
+            try {
+                List<String> params = getPathParameters(key);
+                String ddnPattern = createStructuralNodeRegexString(params, targetUrl, key);
+                if (ddnPattern != null) {
+                    ddnRegexStrings.add(ddnPattern);
+                }
+            } catch (Exception e) {
+                LOG.error("error evaluating DDN for key: {}", key, e);
+            }
+        }
+
+        List<StructuralNodeModifier> structuralNodeModifiers = new ArrayList<>();
+        int ddnCnt = 0;
+        for (String regexString : ddnRegexStrings) {
+            try {
+                Pattern pattern = Pattern.compile(regexString);
+                StructuralNodeModifier structuralNodeModifier =
+                        createStructuralNode(pattern, "DDN" + ddnCnt++);
+                structuralNodeModifiers.add(structuralNodeModifier);
+            } catch (Exception e) {
+                LOG.error("error adding structural node", e);
+            }
+        }
+
+        return structuralNodeModifiers;
+    }
+
+    public void setStructuralNodeModifiers(int contextId) {
+        if (contextId == -1) {
+            return;
+        }
+
+        try {
+            List<StructuralNodeModifier> structuralNodeModifiers = getStructuralNodeModifiers();
+            Context ctx = Model.getSingleton().getSession().getContext(contextId);
+            ctx.setDataDrivenNodes(structuralNodeModifiers);
+        } catch (Exception e) {
+            LOG.error("Error setting data driven nodes from spec for contextId: {}", contextId, e);
+            errors.add(e.getMessage());
+        }
+    }
+
+    private static StructuralNodeModifier createStructuralNode(Pattern pattern, String ddnName) {
+        return new StructuralNodeModifier(
+                StructuralNodeModifier.Type.DataDrivenNode, pattern, ddnName);
+    }
+
+    private static String createStructuralNodeRegexString(
+            List<String> params, String finalTargetUrl, String key) {
+        if (params.isEmpty()) {
+            return null;
+        }
+        // get last param as the DDN for the key
+        String lastParam = params.get(params.size() - 1);
+
+        StringBuilder sb = new StringBuilder();
+        if (finalTargetUrl != null) {
+            sb.append(finalTargetUrl);
+        }
+
+        sb.append('(');
+        String pathTill = key.split(lastParam)[0].replace("{", "");
+        // replace other params with .+? regex in the first path grouping
+        for (String param : params) {
+            if (!param.equals(lastParam)) {
+                pathTill = pathTill.replace(param, ".+?");
+            }
+        }
+        sb.append(pathTill);
+        // add trailing matcher for this DDN
+        sb.append(")(.+?)(/.*)");
+        return removeBrackets(sb.toString());
+    }
+
+    private static List<String> getPathParameters(String key) {
+        String[] pathParts = key.split("/");
+        List<String> params = new ArrayList<>();
+        for (String part : pathParts) {
+            if (PATH_PART_PATTERN.matcher(part).matches()) {
+                params.add(removeBrackets(part));
+            }
+        }
+        return params;
+    }
+
+    private static String removeBrackets(String str) {
+        return str.replace("{", "").replace("}", "");
     }
 }